pyproject.toml
[project]
name = "webp-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "discord-py>=2.5.2",
    "dotenv>=0.9.9",
    "loguru>=0.7.3",
    "requests>=2.32.3",
    "webptools>=0.0.9",
]

[dependency-groups]
dev = [
    "pip>=25.1.1",
    "ruff>=0.11.12",
]
    